infinite scroll behavior added

+++ app/src/main/java/com/mvvm/coroutineexample/ui/main/CharactersAdapter.kt
@@ -12,7 +12,7 @@ class CharactersAdapter(private val listener: CharacterItemListener) :
     RecyclerView.Adapter<CharacterViewHolder>() {
 
     interface CharacterItemListener {
-        fun onClickedCharacter(characterId: Int)
+        fun onClickedCharacter(characterId: Long)
     }
 
     private val items = ArrayList<CharacterModel>()
